Move files which in consensus/moca to consensus/
1. move files which in consensus/moca to consensus/
2. delete moca folder
